requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
torch
torchvision
pyyaml
tqdm